Set up upload module
