+++ package.json
@@ -3,13 +3,12 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "start": "ts-node --files --pretty src/scrape/index.ts",
+    "start": "GIT_BRANCH=`git branch --show-current` ts-node --files --pretty src/index.ts",
     "lint": "scree lint",
     "lint-fix": "scree lint-fix",
     "precommit": "scree precommit",
     "preinstall": "scree preinstall",
-    "install-precommit": "scree install-precommit",
-    "deploy": "AWS_PROFILE=canyoneer GIT_BRANCH=`git branch --show-current` ts-node src/deploy/index.ts"
+    "install-precommit": "scree install-precommit"
   },
   "author": "",
   "license": "ISC",
